# api/openapi.yaml
openapi: 3.0.3
info:
  title: GoChat HTTP API
  version: "0.1.0"
  description: |
    HTTP surface for the GoChat services (auth, contacts, chat view, messaging, presence, and users).
    Endpoints are organised by service; authentication is per-endpoint (Basic for /login, Bearer for protected routes).
servers:
  - url: http://localhost:8080
    description: Default API gateway/edge
paths:
  /me:
    get:
      summary: Get the user id for an access token
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Valid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthInfoResponse'
        '400':
          description: Missing or malformed Authorization header
        '401':
          description: Invalid token
  /login:
    post:
      summary: Exchange username/password for an access token
      security:
        - BasicAuth: []
      responses:
        '200':
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '400':
          description: Authorization header missing or invalid
        '401':
          description: Authentication failed
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid request body
        '500':
          description: Server error while creating the user
  /.well-known/jwks.json:
    get:
      summary: Fetch the JWKS used to validate JWT access tokens
      responses:
        '200':
          description: JWKS payload
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '500':
          description: JWKS not initialised
  /ws:
    get:
      summary: Consume a one-time WebSocket token and return the associated user id
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Token accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WsUseResponse'
        '400':
          description: Token missing
        '401':
          description: Token invalid or expired
    post:
      summary: Create a short-lived WebSocket token for the current user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WsTokenResponse'
        '401':
          description: Authentication failed
  /webauthn/register/begin:
    post:
      summary: Begin WebAuthn registration for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: WebAuthn creation options
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebAuthnCreationOptions'
        '401':
          description: Authentication required
        '500':
          description: Failed to start registration
  /webauthn/register/finish:
    post:
      summary: Finish WebAuthn registration with the credential response
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebAuthnCredentialResponse'
      responses:
        '200':
          description: Passkey stored
        '400':
          description: Invalid WebAuthn response payload
        '401':
          description: Authentication required
        '500':
          description: Failed to store the credential
  /webauthn/login/begin:
    post:
      summary: Begin WebAuthn login
      responses:
        '200':
          description: Assertion options
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebAuthnRequestOptions'
        '500':
          description: Failed to start login
  /webauthn/login/finish:
    post:
      summary: Finish WebAuthn login and obtain an access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebAuthnCredentialResponse'
      responses:
        '200':
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '400':
          description: Invalid assertion payload
        '401':
          description: Authentication failed
  /webauthn/passkeys:
    get:
      summary: List passkeys for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Passkeys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Passkey'
        '401':
          description: Authentication required
  /webauthn/passkeys/{id}:
    put:
      summary: Update a passkey display name
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasskeyRequest'
      responses:
        '200':
          description: Updated passkey
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Passkey'
        '400':
          description: Invalid request
        '401':
          description: Authentication required
        '500':
          description: Update failed
    delete:
      summary: Delete a passkey
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: Deleted
        '400':
          description: Invalid passkey id
        '401':
          description: Authentication required
        '500':
          description: Delete failed
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    BasicAuth:
      type: http
      scheme: basic
...
  schemas:
    AuthInfoResponse:
      type: object
      properties:
        user_id:
          type: integer
          format: int32
      required:
        - user_id
    AuthToken:
      type: object
      properties:
        UserId:
          type: integer
          format: int32
        AccessToken:
          type: string
          description: JWT access token
        RefreshToken:
          type: string
          description: Reserved for future refresh support
        ExpiresIn:
          type: integer
          format: int64
          description: Token lifetime in nanoseconds
        TokenType:
          type: string
      required:
        - UserId
        - AccessToken
        - ExpiresIn
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        name:
          type: string
        password:
          type: string
          format: password
      required:
        - username
        - password
    RegisterResponse:
      type: object
      properties:
        user_id:
          type: integer
          format: int32
      required:
        - user_id
    WsTokenResponse:
      type: object
      properties:
        ws_token:
          type: string
      required:
        - ws_token
    WsUseResponse:
      type: object
      properties:
        user_id:
          type: integer
          format: int32
      required:
        - user_id
    WebAuthnCreationOptions:
      type: object
      description: CredentialCreation object returned by go-webauthn; fields follow the WebAuthn spec.
      additionalProperties: true
    WebAuthnRequestOptions:
      type: object
      description: CredentialRequestOptions for a WebAuthn login.
      additionalProperties: true
    WebAuthnCredentialResponse:
      type: object
      description: Client-side WebAuthn credential response payload.
      additionalProperties: true
    Passkey:
      type: object
      properties:
        id:
          type: integer
          format: int32
        key_name:
          type: string
        user_id:
          type: integer
          format: int32
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
      required:
        - id
        - key_name
        - user_id
        - created_at
    UpdatePasskeyRequest:
      type: object
      properties:
        id:
          type: integer
          format: int32
        key_name:
          type: string
      required:
        - id
        - key_name
...