# pkg/auth/handler.go
package auth

import (
	"fmt"
	"net/http"
	"strconv"
	"strings"

	"github.com/gin-gonic/gin"
	"github.com/go-webauthn/webauthn/protocol"
	"github.com/rs/zerolog"
)

type AuthHandler struct {
	router *gin.Engine
	auth   *AuthService
}

func registerRoutes(r gin.IRouter, h *AuthHandler) {

	r.GET("/me", h.HandleAuthInfo)
	r.POST("/login", h.HandleLogin)
	r.POST("/register", h.HandleRegister)

	r.GET("/.well-known/jwks.json", h.HandleJwks)
	r.GET("/token/exchange", h.HandleTokenExchange)

	r.GET("/ws", h.HandleUseWsToken)
	r.POST("/ws", h.HandleCreateWsToken)

	r.POST("/webauthn/register/begin", h.HandleWebAuthnRegisterBegin)
	r.POST("/webauthn/register/finish", h.HandleWebAuthnRegisterFinish)

	r.POST("/webauthn/login/begin", h.HandleWebAuthnLoginBegin)
	r.POST("webauthn/login/finish", h.HandleWebAuthnLoginFinish)

	r.GET("/webauthn/passkeys", h.HandleGetWebAuthnPasskeys)
	r.PUT("/webauthn/passkeys/:id", h.HandleUpdateWebAuthnPasskey)
	r.DELETE("/webauthn/passkeys/:id", h.HandleDeleteWebAuthnPasskey)
}

func Logger(logger *zerolog.Logger) gin.HandlerFunc {
	return func(c *gin.Context) {
		c.Next()

		logger.Info().
			Str("method", c.Request.Method).
			Str("path", c.Request.URL.Path).
			Int("status", c.Writer.Status()).
			Int("size", c.Writer.Size()).
			Err(c.Errors.Last()).
			Send()
	}
}

func newAuthHandler(authService *AuthService) *AuthHandler {
	router := gin.New()
	router.Use(Logger(authService.logger))
	h := AuthHandler{
		router: router,
		auth:   authService,
	}

	registerRoutes(router, &h)

	return &h
}

func (h *AuthHandler) HandleAuthInfo(c *gin.Context) {
	tokenString, ok := parseBearerToken(c.Request)
	if !ok {
		c.IndentedJSON(http.StatusBadRequest, gin.H{"error": "invalid token"})
		return
	}

	uid, err := h.auth.ValidateAccessToken(c.Request.Context(), tokenString)
	if err != nil {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}
	c.IndentedJSON(http.StatusOK, gin.H{
		"user_id": uid,
	})
}

func (h *AuthHandler) HandleLogin(c *gin.Context) {
	var req struct {
		Username string `json:"username" bind:"required"`
		Password string `json:"password" bind:"required"`
		Client   Client `json:"client,omitempty"`
	}
	err := c.BindJSON(&req)
	if err != nil {
		c.IndentedJSON(http.StatusBadRequest, gin.H{"error": "invalid request body"})
		return
	}
	token, err := h.auth.Login(c.Request.Context(), req.Username, req.Password, &req.Client)
	if err != nil {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}
	c.IndentedJSON(http.StatusOK, token)
}

func (h *AuthHandler) HandleRegister(c *gin.Context) {
	var req struct {
		Username string `json:"username" bind:"required"`
		Name     string `json:"name" bind:"required"`
		Password string `json:"password" bind:"required"`
	}
	err := c.BindJSON(&req)
	if err != nil {
		c.IndentedJSON(http.StatusBadRequest, gin.H{"error": "invalid request body"})
		return
	}
	userId, err := h.auth.RegisterUser(c.Request.Context(), req.Username, req.Password, req.Name)
	if err != nil {
		c.IndentedJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.IndentedJSON(http.StatusOK, gin.H{"user_id": userId})
}

func (h *AuthHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	h.router.ServeHTTP(w, r)
}

func (h *AuthHandler) HandleJwks(c *gin.Context) {
	jwksData, err := h.auth.GetJwksData()
	if err != nil {
		c.IndentedJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.Header("Content-Type", "application/json")
	c.Status(http.StatusOK)
	c.Writer.Write(jwksData)
}

func (h *AuthHandler) HandleTokenExchange(c *gin.Context) {
	var req struct {
		SessionId          string   `json:"session_id" binding:"required"`
		RequestedAudiences []string `json:"request_audiences" binding:"required"`
		RequestedScopes    []string `json:"request_scopes" binding:"required"`
		RequestedTTL       int64    `json:"request_ttl"`
	}
	h.auth.Exchange(c.Request.Context(), req.SessionId, req.RequestedAudiences, req.RequestedScopes, req.RequestedTTL)
}

func (h *AuthHandler) HandleCreateWsToken(c *gin.Context) {
	token, ok := parseBearerToken(c.Request)
	if !ok {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": "invalid token"})
		return
	}
	wsToken, err := h.auth.CreateWsToken(c.Request.Context(), token)
	if err != nil {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}
	c.IndentedJSON(http.StatusOK, gin.H{
		"ws_token": wsToken,
	})
}

func (h *AuthHandler) HandleUseWsToken(c *gin.Context) {
	token := c.Request.URL.Query().Get("token")
	if token == "" {
		c.IndentedJSON(http.StatusBadRequest, gin.H{"error": "missing token query parameter"})
		return
	}
	userId, err := h.auth.UseWsToken(c.Request.Context(), token)
	if err != nil {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}
	c.IndentedJSON(http.StatusOK, gin.H{
		"user_id": userId,
	})
}

func (h *AuthHandler) HandleWebAuthnRegisterBegin(c *gin.Context) {
	accessToken, ok := parseBearerToken(c.Request)
	if !ok {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": "authentication required"})
		return
	}
	uid, err := h.auth.ValidateAccessToken(c.Request.Context(), accessToken)
	if err != nil {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}

	creation, err := h.auth.WebAuthnRegisterBegin(c.Request.Context(), uid)
	if err != nil {
		c.Error(err)
		c.IndentedJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.IndentedJSON(http.StatusOK, creation)
}

func (h *AuthHandler) HandleWebAuthnRegisterFinish(c *gin.Context) {
	accessToken, ok := parseBearerToken(c.Request)
	if !ok {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": "invalid token"})
		return
	}
	userId, err := h.auth.ValidateAccessToken(c.Request.Context(), accessToken)
	if err != nil {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}
	pcc, err := protocol.ParseCredentialCreationResponse(c.Request)
	if err != nil {
		c.IndentedJSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	err = h.auth.WebAuthnRegisterFinish(c.Request.Context(), userId, pcc)
	if err != nil {
		c.IndentedJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
}

func (h *AuthHandler) HandleWebAuthnLoginBegin(c *gin.Context) {
	assertion, err := h.auth.WebAuthnLoginBegin(c.Request.Context())
	if err != nil {
		c.IndentedJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.IndentedJSON(http.StatusOK, assertion)
}

func (h *AuthHandler) HandleWebAuthnLoginFinish(c *gin.Context) {
	pca, err := protocol.ParseCredentialRequestResponse(c.Request)
	if err != nil {
		c.IndentedJSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	token, err := h.auth.WebAuthnLoginFinish(c.Request.Context(), pca, nil)
	if err != nil {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}
	c.IndentedJSON(http.StatusOK, token)
}

func (h *AuthHandler) HandleGetWebAuthnPasskeys(c *gin.Context) {
	accessToken, ok := parseBearerToken(c.Request)
	if !ok {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": "authentication required"})
		return
	}
	uid, err := h.auth.ValidateAccessToken(c.Request.Context(), accessToken)
	if err != nil {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}

	passkeys, err := h.auth.GetWebAuthnUserPasskeys(c.Request.Context(), uid)
	if err != nil {
		c.IndentedJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	fmt.Println(uid)
	c.IndentedJSON(http.StatusOK, passkeys)
}

func (h *AuthHandler) HandleUpdateWebAuthnPasskey(c *gin.Context) {
	accessToken, ok := parseBearerToken(c.Request)
	if !ok {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": "authentication required"})
		return
	}
	uid, err := h.auth.ValidateAccessToken(c.Request.Context(), accessToken)
	if err != nil {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}
	var updateReq struct {
		KeyId   int32  `json:"id" binding:"required"`
		KeyName string `json:"key_name" binding:"required"`
	}
	if err = c.BindJSON(&updateReq); err != nil {
		c.IndentedJSON(http.StatusBadRequest, gin.H{"error": "invalid request body"})
		return
	}
	passkey, err := h.auth.UpdateWebAuthnPasskey(c.Request.Context(), uid, updateReq.KeyId, updateReq.KeyName)
	if err != nil {
		c.IndentedJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.IndentedJSON(http.StatusOK, passkey)
}

func (h *AuthHandler) HandleDeleteWebAuthnPasskey(c *gin.Context) {
	accessToken, ok := parseBearerToken(c.Request)
	if !ok {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": "authentication required"})
		return
	}
	uid, err := h.auth.ValidateAccessToken(c.Request.Context(), accessToken)
	if err != nil {
		c.IndentedJSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}
	keyId, err := parseInt32(c, "id")
	if err != nil {
		c.IndentedJSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	if err = h.auth.DeleteWebAuthnPasskey(c.Request.Context(), uid, keyId); err != nil {
		c.IndentedJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.Status(http.StatusNoContent)
}

func parseUserId(c *gin.Context) (int32, error) {
	uidStr := c.GetString("user_id")
	if uidStr == "" {
		return 0, fmt.Errorf("user_id not found in context")
	}
	uid, err := strconv.ParseInt(uidStr, 10, 32)
	if err != nil {
		return 0, fmt.Errorf("invalid user_id in context")
	}
	return int32(uid), nil
}

func parseBearerToken(r *http.Request) (string, bool) {
	authHeader := r.Header.Get("Authorization")
	parts := strings.Split(authHeader, " ")
	if len(parts) != 2 || parts[0] != "Bearer" {
		return "", false
	}
	return parts[1], true
}

func parseInt32(c *gin.Context, name string) (int32, error) {
	idStr := c.Param(name)
	id, err := strconv.ParseInt(idStr, 10, 32)
	if err != nil {
		return 0, fmt.Errorf("invalid %s: %v", name, err)
	}
	return int32(id), nil
}